+++ src/server/controlchan/handlers/mkd.rs
@@ -8,8 +8,8 @@
 use super::handler::CommandContext;
 use crate::server::chancomms::InternalMsg;
 use crate::server::controlchan::handlers::ControlCommandHandler;
+use crate::server::controlchan::Reply;
 use crate::server::error::FTPError;
-use crate::server::reply::Reply;
 use crate::storage;
 use async_trait::async_trait;
 use futures::channel::mpsc::Sender;
