+++ src/server/controlchan/handlers/pass.rs
@@ -13,9 +13,9 @@
 use super::handler::CommandContext;
 use crate::server::chancomms::InternalMsg;
 use crate::server::controlchan::handlers::ControlCommandHandler;
+use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
 use crate::server::password;
-use crate::server::reply::{Reply, ReplyCode};
 use crate::server::session::SessionState;
 use crate::storage;
 use async_trait::async_trait;
