+++ src/server/controlchan/handlers/rest.rs
@@ -8,8 +8,8 @@
 
 use super::handler::CommandContext;
 use crate::server::controlchan::handlers::ControlCommandHandler;
+use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
-use crate::server::reply::{Reply, ReplyCode};
 use crate::storage;
 use async_trait::async_trait;
