+++ src/server/controlchan/handlers/retr.rs
@@ -8,8 +8,8 @@
 use super::handler::CommandContext;
 use crate::server::controlchan::command::Command;
 use crate::server::controlchan::handlers::ControlCommandHandler;
+use crate::server::controlchan::Reply;
 use crate::server::error::{FTPError, FTPErrorKind};
-use crate::server::reply::Reply;
 use crate::storage;
 use async_trait::async_trait;
 use futures::prelude::*;
