+++ src/server/controlchan/handlers/size.rs
@@ -1,8 +1,8 @@
 use super::handler::CommandContext;
 use crate::server::chancomms::InternalMsg;
 use crate::server::controlchan::handlers::ControlCommandHandler;
+use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
-use crate::server::reply::{Reply, ReplyCode};
 use crate::storage::{self, Metadata};
 use async_trait::async_trait;
 use futures::channel::mpsc::Sender;
