+++ src/server/controlchan/handlers/stat.rs
@@ -20,8 +20,8 @@
 use super::handler::CommandContext;
 use crate::server::chancomms::InternalMsg;
 use crate::server::controlchan::handlers::ControlCommandHandler;
+use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
-use crate::server::reply::{Reply, ReplyCode};
 use crate::storage::{self, Error, ErrorKind};
 use async_trait::async_trait;
 use bytes::Bytes;
