+++ src/server/controlchan/handlers/user.rs
@@ -1,7 +1,7 @@
 use super::handler::CommandContext;
 use crate::server::controlchan::handlers::ControlCommandHandler;
+use crate::server::controlchan::{Reply, ReplyCode};
 use crate::server::error::FTPError;
-use crate::server::reply::{Reply, ReplyCode};
 use crate::server::session::SessionState;
 use crate::storage;
 use async_trait::async_trait;
